Move common classes and functions to pyanaconda.modules.common
The classes and functions that are common to the modules, or need
to be importable by UI, are moved to pyanaconda.modules.common.
For example, base classes for modules and tasks.
