+++ configure.ac
@@ -146,7 +146,6 @@ AC_CONFIG_FILES([Makefile
                  pyanaconda/kickstart_dispatcher/Makefile
                  pyanaconda/payload/Makefile
                  pyanaconda/storage/Makefile
-                 pyanaconda/task/Makefile
                  pyanaconda/ui/Makefile
                  pyanaconda/ui/categories/Makefile
                  pyanaconda/ui/lib/Makefile
@@ -161,6 +160,9 @@ AC_CONFIG_FILES([Makefile
                  pyanaconda/dbus_addons/Makefile
                  pyanaconda/dbus_addons/baz/Makefile
                  pyanaconda/modules/Makefile
+                 pyanaconda/modules/common/Makefile
+                 pyanaconda/modules/common/base/Makefile
+                 pyanaconda/modules/common/task/Makefile
                  pyanaconda/modules/bar/Makefile
                  pyanaconda/modules/bar/tasks/Makefile
                  pyanaconda/modules/boss/Makefile
