+++ pyanaconda/modules/common/Makefile.am
@@ -1,6 +1,5 @@
-# modules/Makefile.am for anaconda
 #
-# Copyright (C) 2017  Red Hat, Inc.
+# Copyright (C) 2018  Red Hat, Inc.
 #
 # This program is free software; you can redistribute it and/or modify
 # it under the terms of the GNU Lesser General Public License as published
@@ -15,10 +14,10 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-SUBDIRS = common boss bar foo timezone network
+SUBDIRS = base task
 
 pkgpyexecdir = $(pyexecdir)/py$(PACKAGE_NAME)
-modulesdir = $(pkgpyexecdir)/modules
-modules_PYTHON = $(srcdir)/*.py
+modules_commondir = $(pkgpyexecdir)/modules/common
+modules_common_PYTHON = $(srcdir)/*.py
 
 MAINTAINERCLEANFILES = Makefile.in
