+++ pyanaconda/modules/common/base/base.py
@@ -24,7 +24,7 @@ from pyanaconda.core.event_loop import EventLoop
 from pyanaconda.core.async_utils import run_in_loop
 from pyanaconda.core.timer import Timer
 from pyanaconda.dbus import DBus
-from pyanaconda.task import publish_task
+from pyanaconda.modules.common.task import publish_task
 from pyanaconda.core.signal import Signal
 from pyanaconda.core.kickstart import NoKickstartSpecification, \
     KickstartSpecificationHandler, KickstartSpecificationParser
