+++ pyanaconda/modules/common/task/task.py
@@ -26,11 +26,11 @@ from abc import ABC, abstractmethod
 
 from pyanaconda.core.signal import Signal
 from pyanaconda.core.constants import THREAD_DBUS_TASK
-from pyanaconda.task.task_interface import TaskAlreadyRunningException
+from pyanaconda.modules.common.task.task_interface import TaskAlreadyRunningException
 from pyanaconda.threading import threadMgr, AnacondaThread
 from pyanaconda.core.async_utils import async_action_nowait
 
-__all__ = ['Task', 'TaskAlreadyRunningException']
+__all__ = ['Task']
 
 
 class Task(ABC):
