+++ pyanaconda/modules/foo/foo_interface.py
@@ -19,7 +19,7 @@
 #
 
 from pyanaconda.dbus.constants import MODULE_FOO_NAME
-from pyanaconda.modules.base_interface import KickstartModuleInterface
+from pyanaconda.modules.common.base import KickstartModuleInterface
 from pyanaconda.dbus.interface import dbus_interface
