+++ pyanaconda/modules/timezone/timezone_interface.py
@@ -21,7 +21,7 @@
 from pyanaconda.dbus.constants import MODULE_TIMEZONE_NAME
 from pyanaconda.dbus.property import emits_properties_changed
 from pyanaconda.dbus.typing import *  # pylint: disable=wildcard-import
-from pyanaconda.modules.base_interface import KickstartModuleInterface
+from pyanaconda.modules.common.base import KickstartModuleInterface
 from pyanaconda.dbus.interface import dbus_interface
